# EnumSet: added getEnumerators(), getNames() and getValues()