see #62 Add methods to convert a set into an array

* `EnumSet::getEnumerators() : Enum[]`
* `EnumSet::getValues() : scalar[]`
* `EnumSet::getNames() : string[]`

@prolic I decided to use this names to be consistent with the already existing method `Enum::getEnumerators()` and I also return it just as lists so the ordinal number will not be the array key. I analyzed all my use-cases and couldn't find any where the ordinal number as array key is required. On the other hand I have some use-cases where I actually require a simple list. Also it's still simple possible to just use `iterator_to_array($set)` to convert the set into an associative array with ordinal number to enumerator instance.

PS: The same set of methods should also work on `EnumMap` and on `Enum` (as static versions)