@marc-mabe 2016-09-30
I did some changes to optimize the performance of the methods and to not reset the internal iterator position ;)

Also added the method `getOrdinals() : int[]` as this offers the basic information stored in a set as array and all other `get*s()` methods are based on that.

## Benchmark

Script: https://gist.github.com/marc-mabe/a26227adf5ea1e9c96c2072bf4351ec1
Run on:
* PHP 7.0.8-0ubuntu0.16.04.2 (cli) ( NTS )
* Intel(R) Core(TM) i7 CPU         920  @ 2.67GHz

### Before:
```
$ php bench.php 
empty set getEnumerators(): 0.0085999965667725ms
empty set getValues()     : 0.0090000629425049ms
empty set getNames()      : 0.0085229873657227ms
full set getEnumerators() : 1.471755027771ms
full set getValues()      : 1.533252954483ms
full set getNames()       : 1.9099199771881ms
```

### After:
```
$ php bench.php 
empty set getEnumerators(): 0.0077300071716309ms
empty set getValues()     : 0.0088911056518555ms
empty set getNames()      : 0.0087888240814209ms
full set getEnumerators() : 0.68992781639099ms
full set getValues()      : 0.74091100692749ms
full set getNames()       : 1.0174210071564ms
```