@marc-mabe 2016-09-30
@prolic thanks for review :)